requirements.txt
# Core dependencies for The Librarian production deployment
openai>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
letta-client>=0.1.178
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
tiktoken>=0.5.0

# Optional dependencies for enhanced features
# redis>=4.0.0  # For caching
# prometheus-client>=0.15.0  # For metrics
